requirements.txt
groq>=0.8.0
python-dotenv>=1.0.0
requests>=2.31.0
matplotlib>=3.8.0
numpy>=1.26.0
 